Refactor tasklist-class.js file by moving some methods to separate files in order to have more modules to test
