--- src/CompletedFilter.js
@@ -1,3 +1,0 @@
-const completedIsFalse = (task) => task.completed === false;
-
-export default completedIsFalse;
