--- src/TaskObject.js
@@ -1,9 +1,0 @@
-class TaskObject {
-  constructor(description, index) {
-    this.description = description;
-    this.completed = false;
-    this.index = index + 1;
-  }
-}
-
-export default TaskObject;
